This commit was manufactured by cvs2svn to create tag 'SDK_1-2_PreMerge'.
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/tags/SDK_1-2_PreMerge@209 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ game/playerview.cpp
@@ -7,8 +7,14 @@
  * $Author$
  *
  * $Log$
- * Revision 1.1  2004/10/30 15:52:31  sparhawk
- * Initial revision
+ * Revision 1.3  2005/10/18 13:56:41  sparhawk
+ * Lightgem updates
+ *
+ * Revision 1.2  2004/10/30 16:55:57  sparhawk
+ * Compile errors fixed. float cast added.
+ *
+ * Revision 1.1.1.1  2004/10/30 15:52:31  sparhawk
+ * Initial release
  *
  ***************************************************************************/
 
@@ -537,7 +543,7 @@ void idPlayerView::DoubleVision( idUserInterface *hud, const renderView_t *view,
 	if ( scale > 0.5f ) {
 		scale = 0.5f;
 	}
-	float shift = scale * sin( sqrt( offset ) * g_dvFrequency.GetFloat() );
+	float shift = scale * sin( sqrt( (float)offset ) * g_dvFrequency.GetFloat() );
 	shift = fabs( shift );
 
 	// if double vision, render to a texture
@@ -673,7 +679,7 @@ void idPlayerView::InfluenceVision( idUserInterface *hud, const renderView_t *vi
 		SingleView( hud, view );
 		return;
 	} else {
-		int offset =  25 + sin( gameLocal.time );
+		int offset =  25 + sin( (float)gameLocal.time );
 		DoubleVision( hud, view, pct * offset );
 	}
 }
@@ -683,12 +689,16 @@ void idPlayerView::InfluenceVision( idUserInterface *hud, const renderView_t *vi
 idPlayerView::RenderPlayerView
 ===================
 */
-void idPlayerView::RenderPlayerView( idUserInterface *hud ) {
+void idPlayerView::RenderPlayerView( idUserInterface *hud )
+{
 	const renderView_t *view = player->GetRenderView();
 
-	if ( g_skipViewEffects.GetBool() ) {
+	if(g_skipViewEffects.GetBool())
+	{
 		SingleView( hud, view );
-	} else {
+	}
+	else 
+	{
 		if ( player->GetInfluenceMaterial() || player->GetInfluenceEntity() ) {
 			InfluenceVision( hud, view );
 		} else if ( gameLocal.time < dvFinishTime ) {
@@ -701,4 +711,5 @@ void idPlayerView::RenderPlayerView( idUserInterface *hud ) {
 		ScreenFade();
 	}
 }
+
